arch: move MSR functions to arch/x86
